backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# Google APIs
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.116.0
google-cloud-texttospeech>=2.16.3

# LangChain & Gemini (updated for Python 3.13)
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-core>=0.1.0
langgraph>=0.0.20

# Deepgram
deepgram-sdk==3.2.7

# Utilities (updated for Python 3.13 compatibility)
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
pytz>=2023.3
python-dateutil>=2.8.2
certifi>=2023.7.22

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

